Make it possible to add tools (#14)
Makes it simpler to add more tools, or use none. We'll use this to add
Go and TS language server tools.

The pydantic-ai classes are nice here to do the mapping between
functions and json schemas. There's a bit of weirdness to use them
outside of a full pydantic-ai agent, but it isn't too bad.